pyproject.toml
[project]
name = "ratatouille-royale"
version = "0.1.0"
description = ""
authors = [
    {name = "WingedSeal", email = "wingedseal@iusevimbtw.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce (==2.5.5)",
    "pygame-gui (==0.6.14)",
    "python-i18n (==0.3.9)",
    "pytweening (==1.2.0)"
]

[project.scripts]
ratroyale = "main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "numpy (>=2.3.3,<3.0.0)",
    "poethepoet (>=0.37.0,<0.38.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "isort (>=6.1.0,<7.0.0)"
]

[tool.poetry]
packages = [
    {include = "ratroyale", from = "src"},
    {include = "main.py", from = "src"}
]

[tool.poe.tasks]
render-test = "python ./src/render_test.py"
format = "black ./src ./tests"
check = "pre-commit run --all-files"
tiled-cli = { cmd = "python ${POE_ROOT}/src/tiled_cli.py", cwd = "${POE_PWD}" }

[tool.mypy]
python_version = "3.12"
mypy_path = ["src", "tests"]
files = ["./src/ratroyale/", "./src/main.py", "./tests/"]
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.black]
line-length = 88
target-version = ["py312"]
include = "\\.pyi?$"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [
  "src"
]
markers = [
    "slow: slow running tests",
    "integration: integration tests", 
    "backend: backend tests",
]
